Разделил mainScreen на несколько файлов, немного изменил console и settings
