+++ app/src/main/java/com/example/hits_android/Console.kt
@@ -1,0 +1,48 @@
+package com.example.hits_android
+
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.fillMaxHeight
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.lazy.LazyColumn
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.collectAsState
+import androidx.compose.runtime.getValue
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.unit.dp
+import androidx.compose.ui.unit.sp
+import com.example.hits_android.model.FlowViewModel
+
+@Composable
+fun Console(viewModel: FlowViewModel) {
+    val output by viewModel.output.collectAsState()
+
+    LazyColumn(
+        modifier = Modifier.fillMaxWidth().fillMaxHeight(0.93f)
+    ) {
+        item {
+            Box(
+                modifier = Modifier.fillMaxWidth(),
+                contentAlignment = Alignment.Center
+            ) {
+                Text(
+                    text = "Console",
+                    fontSize = 30.sp,
+                    fontWeight = FontWeight.Bold,
+                    modifier = Modifier.padding(15.dp)
+                )
+
+            }
+        }
+        item {
+            Text(
+                text = output,
+                modifier = Modifier.padding(15.dp)
+            )
+        }
+    }
+}
