+++ app/src/main/java/com/example/hits_android/MainScreen.kt
@@ -85,7 +85,7 @@ fun MainScreen(
 ) {
     val theme by vm.theme.collectAsState()
     Hits_androidTheme(buildTheme(theme)) {
-        BottomNav(vm = vm)
+        BottomNavigation(vm = vm)
     }
 }
 
@@ -99,556 +99,5 @@ fun Sandbox(
             .fillMaxWidth()
     ) {
         VerticalReorderList(vm = vm)
-    }
-}
-
-
-@OptIn(ExperimentalMaterial3Api::class)
-@Composable
-private fun VerticalReorderList(
-    vm: ReorderListViewModel
-) {
-    val state = rememberReorderableLazyListState(
-        onMove = vm::moveBlock,
-        canDragOver = vm::isDogDragOverEnabled
-    )
-    LazyColumn(
-        state = state.listState,
-        modifier = Modifier.reorderable(state).padding(start = 4.dp, end = 4.dp),
-        verticalArrangement = Arrangement.spacedBy(8.dp)
-    ) {
-        items(vm.codeBlocksList, { item -> item.key }) { item ->
-            ReorderableItem(state, item.key) { dragging ->
-                val scale = animateFloatAsState(if (dragging) 1.1f else 1.0f)
-                val elevation = animateDpAsState(if (dragging) 8.dp else 0.dp)
-                val dismissState = rememberDismissState(
-                    confirmValueChange = {
-                        if ((it == DismissValue.DismissedToEnd || it == DismissValue.DismissedToStart)) {
-                            if (item.isDragOverLocked) {
-                                vm.codeBlocksList = vm.codeBlocksList.toMutableList().apply {
-                                    removeIf { it.key != "0" && it.key != "1" }
-                                }
-                                return@rememberDismissState false
-                            } else if (item.blockName == "endBlock" || item.blockName == "beginBlock") {
-                                return@rememberDismissState false
-                            } else {
-                                if (item.blockName == "ElseBlock" || item.blockName == "IfBlock" || item.blockName == "WhileBlock") {
-                                    when (item.blockName) {
-                                        "ElseBlock" -> {
-                                            item as ElseBlock
-                                            vm.codeBlocksList =
-                                                vm.codeBlocksList.toMutableList().apply {
-                                                    removeIf { it.key == item.key || it.key == item.beginKey || it.key == item.endKey }
-                                                }
-                                        }
-
-                                        "IfBlock" -> {
-                                            item as IfBlock
-                                            vm.codeBlocksList =
-                                                vm.codeBlocksList.toMutableList().apply {
-                                                    removeIf { it.key == item.key || it.key == item.beginKey || it.key == item.endKey }
-                                                }
-                                        }
-
-                                        "WhileBlock" -> {
-                                            item as WhileBlock
-                                            vm.codeBlocksList =
-                                                vm.codeBlocksList.toMutableList().apply {
-                                                    removeIf { it.key == item.key || it.key == item.beginKey || it.key == item.endKey }
-                                                }
-                                        }
-                                    }
-
-                                } else {
-                                    vm.codeBlocksList = vm.codeBlocksList.toMutableList().apply {
-                                        removeIf { it.key == item.key }
-                                    }
-                                }
-                                blockList = vm.codeBlocksList.toMutableList()
-                            }
-                        }
-                        true
-                    }
-                )
-                SwipeToDismiss(
-                    state = dismissState,
-                    background = {
-                        val direction = dismissState.dismissDirection ?: return@SwipeToDismiss
-                        val color by animateColorAsState(
-                            targetValue =
-                            when (dismissState.targetValue) {
-                                DismissValue.Default -> MaterialTheme.colorScheme.errorContainer.copy(
-                                    0.1f
-                                )
-
-                                DismissValue.DismissedToEnd -> MaterialTheme.colorScheme.errorContainer
-                                DismissValue.DismissedToStart -> MaterialTheme.colorScheme.errorContainer
-                            }
-                        )
-                        val icon = Icons.Default.Delete
-
-                        if (!(item.isDragOverLocked || item.blockName == "endBlock" || item.blockName == "beginBlock")) {
-                            Box(
-                                modifier = Modifier
-                                    .fillMaxSize()
-                                    .background(
-                                        brush =
-                                        when (direction) {
-                                            DismissDirection.StartToEnd -> Brush.horizontalGradient(
-                                                colors = listOf(color.copy(0.6f), color.copy(0f)),
-                                                startX = 0f,
-                                                endX = 300f
-                                            )
-
-                                            DismissDirection.EndToStart -> Brush.horizontalGradient(
-                                                colors = listOf(color.copy(0.6f), color.copy(0f)),
-                                                startX = 1000f,
-                                                endX = 700f
-                                            )
-                                        }
-                                    )
-                                    .padding(horizontal = 25.dp),
-                                contentAlignment = Alignment.CenterStart
-                            ) {
-                                Icon(
-                                    icon,
-                                    contentDescription = "deleteIcon",
-                                    tint = MaterialTheme.colorScheme.onErrorContainer,
-                                    modifier = Modifier.align(
-                                        when (direction) {
-                                            DismissDirection.StartToEnd -> Alignment.CenterStart
-                                            DismissDirection.EndToStart -> Alignment.CenterEnd
-                                        }
-                                    )
-                                )
-                            }
-                        }
-                    },
-                    directions = setOf(DismissDirection.EndToStart, DismissDirection.StartToEnd),
-                    dismissContent = {
-                        if (item.isDragOverLocked) {
-                            Box(
-                                contentAlignment = Alignment.Center,
-                                modifier = Modifier
-                                    .fillMaxSize()
-                                    .scale(scale.value)
-                                    .shadow(elevation.value, RoundedCornerShape(24.dp))
-                            ) {
-                                item.BlockComposable(item, vm.codeBlocksList)
-                            }
-                        } else {
-                            Box(
-                                contentAlignment = Alignment.Center,
-                                modifier = Modifier
-                                    .fillMaxSize()
-                                    .detectReorderAfterLongPress(state)
-                                    .scale(scale.value)
-                                    .shadow(elevation.value, RoundedCornerShape(24.dp))
-                                    .clickable(
-                                        onClick = {
-                                            Log.d("a", "${vm.codeBlocksList}")
-                                        }
-                                    )
-                            ) {
-                                item.BlockComposable(item, vm.codeBlocksList)
-                            }
-                        }
-                    }
-                )
-            }
-        }
-    }
-}
-
-@Composable
-private fun BottomBar(
-    vm: ReorderListViewModel
-) {
-    LazyRow(
-        horizontalArrangement = Arrangement.spacedBy(8.dp)
-    ) {
-        items(vm.blockSelectionList, { item -> item.key }) { item ->
-            Box(
-                contentAlignment = Alignment.Center,
-                modifier = Modifier
-                    .size(100.dp, 75.dp)
-                    .clip(RoundedCornerShape(24.dp))
-                    .background(MaterialTheme.colorScheme.surfaceVariant)
-                    .clickable(
-                        onClick = {
-                            vm.addBlock(item)
-                        }
-                    )
-            ) {
-                Text(
-                    text = item.title,
-                    color = MaterialTheme.colorScheme.onSurfaceVariant,
-                    fontWeight = FontWeight.Bold
-                )
-            }
-        }
-    }
-}
-
-sealed class BottomBarScreen(
-    val route: String,
-    val title: String,
-    val icon: Int = R.drawable.ic_bottom_home,
-    val icon_focused: Int = R.drawable.ic_bottom_home_focused
-) {
-
-    object BlocksBar : BottomBarScreen(
-        route = "blocksbar",
-        title = "BlocksBar",
-        icon = R.drawable.ic_bottom_blocks_open,
-        icon_focused = R.drawable.ic_bottom_blocks_close
-    )
-
-    object Coding : BottomBarScreen(
-        route = "coding",
-        title = "Coding",
-        icon = R.drawable.ic_bottom_home,
-        icon_focused = R.drawable.ic_bottom_home_focused
-    )
-
-    object Console : BottomBarScreen(
-        route = "console",
-        title = "Console",
-        icon = R.drawable.ic_bottom_report,
-        icon_focused = R.drawable.ic_bottom_report_focused
-    )
-
-    object Settings : BottomBarScreen(
-        route = "settings",
-        title = "Settings",
-        icon = R.drawable.ic_bottom_profile,
-        icon_focused = R.drawable.ic_bottom_profile_focused
-    )
-
-    object Start : BottomBarScreen(
-        route = "start",
-        title = "Start",
-        icon = R.drawable.ic_bottom_play,
-        icon_focused = R.drawable.ic_bottom_stop
-    )
-}
-
-@Composable
-fun BottomNav(
-    vm: ReorderListViewModel
-) {
-    val navController = rememberNavController()
-
-    Scaffold(
-        bottomBar = {
-            NavBottomBar(
-                navController = navController,
-                vm = vm
-            )
-        }
-    ) {
-        Modifier.padding(it)
-        BottomNavGraph(
-            navController = navController,
-            vm = vm
-        )
-    }
-}
-
-@Composable
-fun NavBottomBar(
-    navController: NavHostController,
-    vm: ReorderListViewModel
-) {
-    val viewModel: FlowViewModel = viewModel()
-
-    val screens = listOf(
-        BottomBarScreen.BlocksBar,
-        BottomBarScreen.Coding,
-        BottomBarScreen.Console,
-        BottomBarScreen.Settings,
-        BottomBarScreen.Start
-    )
-
-    val navStackBackEntry by navController.currentBackStackEntryAsState()
-    val currentDestination = navStackBackEntry?.destination
-
-    Row(
-        modifier = Modifier
-            .padding(start = 10.dp, end = 10.dp, top = 8.dp, bottom = 8.dp)
-            .background(Color.Transparent)
-            .fillMaxWidth(),
-        horizontalArrangement = Arrangement.SpaceEvenly,
-        verticalAlignment = Alignment.CenterVertically
-    ) {
-        screens.forEach { screen ->
-            AddItem(
-                screen = screen,
-                currentDestination = currentDestination,
-                navController = navController,
-                viewModel = viewModel,
-                vm = vm
-            )
-        }
-    }
-}
-
-@Composable
-fun BottomNavGraph(
-    navController: NavHostController,
-    vm: ReorderListViewModel
-) {
-
-    val viewModel: FlowViewModel = viewModel()
-    NavHost(
-        navController = navController,
-        startDestination = BottomBarScreen.Coding.route
-    ) {
-        composable(route = BottomBarScreen.Coding.route) {
-            Surface {
-                Column {
-                    Sandbox(vm)
-                }
-            }
-        }
-        composable(route = BottomBarScreen.Console.route) {
-            Console(viewModel)
-        }
-        composable(route = BottomBarScreen.Settings.route) {
-            Settings(vm)
-        }
-    }
-}
-
-@Composable
-fun RowScope.AddItem(
-    screen: BottomBarScreen,
-    currentDestination: NavDestination?,
-    navController: NavHostController,
-    viewModel: FlowViewModel,
-    vm: ReorderListViewModel
-) {
-    val isProgramRunning by viewModel.isProgramRunning.collectAsState()
-    val selected = currentDestination?.hierarchy?.any { it.route == screen.route } == true
-    val theme by vm.theme.collectAsState()
-
-    val background =
-        if (selected) MaterialTheme.colorScheme.primary.copy(alpha = 0.8f) else Color.Transparent
-
-    val contentColor =
-        when (theme.first) {
-            AppThemeBrightness.Light -> if (selected) Color.White else Color.Black
-            AppThemeBrightness.Dark -> Color.White
-            AppThemeBrightness.System -> if (isSystemInDarkTheme()) Color.White else (if (selected) Color.White else Color.Black)
-        }
-
-    val buttonIcon = if (screen == BottomBarScreen.Start) {
-        if (isProgramRunning) R.drawable.ic_bottom_stop else R.drawable.ic_bottom_play
-    } else {
-        if (selected) screen.icon_focused else screen.icon
-    }
-
-    val blocksBarSelected = remember { mutableStateOf(false) }
-
-    Box(
-        modifier = Modifier
-            .height(40.dp)
-            .clip(CircleShape)
-            .background(background)
-            .clickable(onClick = {
-                if (screen == BottomBarScreen.Start) {
-                    if (isProgramRunning) {
-                        viewModel.stopProgram()
-                    } else {
-                        viewModel.startProgram()
-
-                        if (currentDestination?.route != BottomBarScreen.Console.route) {
-                            navController.navigate(BottomBarScreen.Console.route)
-                        }
-                    }
-                } else if (screen == BottomBarScreen.BlocksBar) {
-                    blocksBarSelected.value = !blocksBarSelected.value
-
-                    if (currentDestination?.route != BottomBarScreen.Coding.route) {
-                        navController.navigate(BottomBarScreen.Coding.route)
-                    }
-                } else {
-                    navController.navigate(screen.route) {
-                        popUpTo(navController.graph.findStartDestination().id)
-                        launchSingleTop = true
-                    }
-                }
-            })
-    ) {
-        Row(
-            modifier = Modifier
-                .padding(start = 10.dp, end = 10.dp, top = 8.dp, bottom = 8.dp),
-            verticalAlignment = Alignment.CenterVertically,
-            horizontalArrangement = Arrangement.spacedBy(4.dp)
-        ) {
-            Icon(
-                painter = painterResource(id = if (screen == BottomBarScreen.BlocksBar && blocksBarSelected.value) screen.icon_focused else buttonIcon),
-                contentDescription = "icon",
-                tint = contentColor,
-                modifier = Modifier.size(24.dp)
-            )
-            AnimatedVisibility(visible = blocksBarSelected.value && screen == BottomBarScreen.BlocksBar) {
-                BottomBar(vm = vm)
-            }
-            AnimatedVisibility(visible = selected && screen != BottomBarScreen.Start && screen != BottomBarScreen.BlocksBar) {
-                Text(
-                    text = screen.title,
-                    color = contentColor
-                )
-            }
-        }
-    }
-}
-
-
-@Composable
-fun Console(viewModel: FlowViewModel) {
-    val output by viewModel.output.collectAsState()
-
-    LazyColumn(
-        modifier = Modifier.fillMaxSize()
-    ) {
-        item {
-            Column(
-                modifier = Modifier.fillMaxWidth(),
-                verticalArrangement = Arrangement.Top,
-                horizontalAlignment = Alignment.CenterHorizontally
-            ) {
-                Text(
-                    text = "Console",
-                    fontSize = 20.sp,
-                    fontWeight = FontWeight.Bold
-                )
-
-            }
-        }
-        item {
-            Text(
-                text = output,
-                modifier = Modifier.padding(16.dp)
-            )
-        }
-    }
-}
-
-@Composable
-fun Settings(vm: ReorderListViewModel) {
-
-    val theme by vm.theme.collectAsState()
-    LazyColumn(
-        modifier = Modifier.fillMaxSize()
-    ) {
-        item {
-            Column(
-                modifier = Modifier.fillMaxWidth(),
-                verticalArrangement = Arrangement.Top,
-                horizontalAlignment = Alignment.CenterHorizontally
-            ) {
-                Text(
-                    text = "Settings",
-                    fontSize = 20.sp,
-                    fontWeight = FontWeight.Bold
-                )
-
-            }
-        }
-        item {
-            Row(
-                horizontalArrangement = Arrangement.spacedBy(10.dp),
-                modifier = Modifier
-                    .fillMaxWidth()
-                    .padding(25.dp),
-                verticalAlignment = Alignment.CenterVertically
-            ) {
-                Text(
-                    text = "Theme: ",
-                    fontSize = 15.sp,
-                    fontWeight = FontWeight.Bold,
-                    color = MaterialTheme.colorScheme.onBackground
-                )
-                Box(
-                    modifier = Modifier
-                        .height(70.dp)
-                        .fillMaxWidth(0.5f)
-                        .clip(shape = RoundedCornerShape(24.dp))
-                        .background(MaterialTheme.colorScheme.surfaceVariant)
-                        .clickable {
-                            vm.setCurrentTheme(
-                                Pair(
-                                    when (theme.first) {
-                                        AppThemeBrightness.Light -> AppThemeBrightness.Dark
-                                        AppThemeBrightness.Dark -> AppThemeBrightness.System
-                                        AppThemeBrightness.System -> AppThemeBrightness.Light
-                                    },
-                                    theme.second
-                                )
-                            )
-                        },
-                    contentAlignment = Alignment.Center
-                ) {
-                    Text(
-                        text = when (theme.first) {
-                            AppThemeBrightness.Light -> "Light"
-                            AppThemeBrightness.Dark -> "Dark"
-                            AppThemeBrightness.System -> "System"
-
-                        },
-                        color = MaterialTheme.colorScheme.onSurfaceVariant,
-                        fontWeight = FontWeight.Bold,
-                        fontSize = 30.sp,
-                        maxLines = 1
-                    )
-                }
-                Box(
-                    modifier = Modifier
-                        .height(70.dp)
-                        .fillMaxWidth()
-                        .clip(shape = RoundedCornerShape(24.dp))
-                        .background(MaterialTheme.colorScheme.primary)
-                        .clickable {
-                            vm.setCurrentTheme(
-                                Pair(
-                                    theme.first,
-                                    when (theme.second) {
-                                        AppThemeColor.Green -> AppThemeColor.Purple
-                                        AppThemeColor.Purple -> AppThemeColor.Pink
-                                        AppThemeColor.Pink -> AppThemeColor.Blue
-                                        AppThemeColor.Blue -> AppThemeColor.Red
-                                        AppThemeColor.Red -> AppThemeColor.Yellow
-                                        AppThemeColor.Yellow -> AppThemeColor.Green
-                                    }
-                                )
-                            )
-                        },
-                    contentAlignment = Alignment.Center
-                ) {
-                    Text(
-                        text = when (theme.second) {
-                            AppThemeColor.Green -> "Green"
-                            AppThemeColor.Purple -> "Purple"
-                            AppThemeColor.Pink -> "Pink"
-                            AppThemeColor.Blue -> "Blue"
-                            AppThemeColor.Red -> "Red"
-                            AppThemeColor.Yellow -> "Yellow"
-                        },
-                        color = MaterialTheme.colorScheme.onPrimary,
-                        fontWeight = FontWeight.Bold,
-                        fontSize = 30.sp,
-                        maxLines = 1
-                    )
-                }
-            }
-        }
-        item {
-            Text(
-                text = "Авторы:\nАлексей Ковкин\nЮрий Ситдиков\nАлександр Кирсанов",
-                modifier = Modifier.padding(16.dp),
-            )
-        }
     }
 }
